Installation: minor correction in links to system libraries
Until now when making a link to the system's 'dl' and 'pthread' libraries
we were simply linking the installed location on the system (in
'/usr/lib'). However, in some systems, these may themselves be links to
other locations and this could cause linking problems.

With this commit, we now use 'realpath' to extract the absolute address of
the final file that the libraries may link to, and directly link to them.

A minor cosmetic correction was also made in the build rule for CFITSIO:
the long line was broken into two!