Configuration: --debug option available in this phase also
Until now, each time there was a problem in the configuration of Maneage'd
projects and debugging was necessary, we had to take the following changes:

 - Run the configuration on a single thread ('-j1') to see the building of
   only the problematic software.

 - Disable the Zenodo check manually by commenting those parts of
   'reproduce/software/shell/configure.sh'. Because the internet connection
   wastes a few seconds and is thus very annoying during repeated runs!

 - Manually remove the '-k' option that was passed to Make (when building
   the software). With the '-k', Make keeps going with the execution of
   other targets if something crashes and this usually causes confusions
   during the debugging.

Doing the manual changes within the code was both very annoying and prone
to errors (forgetting to correct it!).

With this commit, the existing '--debug' option has been generalized to the
software configuration phase of Maneage also. Until now, it was only
available in the analysis phase (and would directly be passed to the 'make'
command that would run the analysis). When this option is used, and the
project is in the software configuration phase, the Zenodo check won't be
done, it will use one single thread ('-j1'), and it will stop the execution
as soon as an error occurs (Make is not run with '-k').