./project: unused --minmapsize option is removed
Until now, the './project' script included an '--minmapsize' option which
is an option to one of the original programs that was used in Maneage
(Gnuastro). Such an option doesn't exist in many other programs, so it is
not a suitable option for the generic Maneage project (and can just cause
confusion). It was also not used in any part of Maneage any more!

With this commit, this option is removed from the core Maneage './project'
script and if any project uses it, they can implement it in their own
branch.