Update publications: Peper+Roukema published
This commit updates some of the publication data
in README-hacking.md : Peper+Roukema (2021) is now
published in MNRAS and Akhlaghi+ (2021) is published online and
very close to getting a conventional volume and page number. :)

See task
  https://savannah.nongnu.org/task/?15736
for ideas of how to make a more systematic publication
list instead one managed by prose text. There
are already too many non-automated places for publication
lists where we have to copy/paste our publication data again
and again and again and ...

This commit also adds the softwareheritage ID that we have in the
content of Akhlaghi+2021 (without the extra context, because as a
URL that's very long). There are plenty of arguments to be made
each way for different versions of the swh IDS. One advantage of
the 'rev' ID is that the hash is the original (full) git hash,
which is what I've done for the elaphrocentre and subpoisson
papers.