Configuration: fixed bugs in building of OpenSSL and Gettext
Until now, the 'RPATH' variable (specifying where to look for shared
libraries) wasn't being set in the 'libcrypto' library of OpenSSL (it was
only set for the 'libssl' library).

Also, Gettext used the host Emacs for some operations during installation
that could cause the following crash (because we are giving priority to
local libraries, which the host Emacs doesn't recognize):

  emacs: /BDIR/libcrypto.so.1.1: version `OPENSSL_1_1_1b' not found
  (required by /lib64/libk5crypto.so.3)

With this commit both these bugs have been fixed: 1) Patchelf is run on the
'libcrypto' library also and 2) we pass the '--without-emacs' configuration
option to the configure script of Gettext.

These bugs were found by Elham Saremi.