initialize.mk: Git call in variable works with LD_LIBRARY_PATH
Until now, the '$(project-commit-hash)' Make variable of 'initialize.mk'
simply called 'git' to find the commit hash. However, due to one of the
recent software updates, we noticed that this command is no longer working
(and the project commit hash wasn't getting printed in the PDF)!  The
problem was that Maneage's Git, couldn't find the 'libiconv' library that
it was built with.

With this commit, the '$(shell' command that calls Git, first exports
'LD_LIBRARY_PATH' to Maneage's software build directory. As a result, the
Git command can work and will report the commit as a LaTeX macro to be used
in the paper. To avoid relying on PATH outside of Make recipes, we now also
directly call the Git executable with Maneage.

Some other minor issues have been found and fixed in this commit:

 - README-hacking.md: some minor edits and typo corrections.

 - initialize.mk: the '$(curdir)' variable is now used in several places
   that we were calling 'pwd'.

 - versions.conf: 'xlsxio-version' now included with other programs. Until
   now it was commented because GCC 11.1.0 had issues with it. However, GCC
   11.2.0 doesn't have a problem any more, so it has been returned to the
   list of all high-level programs.

 - xorg.mk: used same format to comment recipe lines as the other Makefiles
   (a '#' followed by a TAB).

 - preamble-pgfplots.tex: lines to comment for building an EPS figure with
   PGFPlots have been re-formatted to be more human-readable.