Added compiler flag necessary to build the `file` program on CentOS 7.
Until now, Maneage failed to build the `file` target on at least one machine running
Centos version 7.9.2009 with kernel 3.10.0 when running `./project configure`.

As suggested by the error message issued by the compiler, this commmits adds
the CFLAG "-std=c99" to the `file` recipe within `basic.mk`.
This flag puts the compiler in C99 mode, which forces it to compile code
according to the 1999 edition of the C standard. With this fix, the target
compiles succesfully.